tests/integration/gateway-vs-actors/tests/requirements.txt
# Dependencies for test runner
pytest>=8.0.0
requests>=2.31.0
sseclient-py>=1.8.0
